Make CircuitCreateTemplate the entrance point to the API
Before the API user would load an YAML template and directly convert it into builders to create circuits via
`Builders::try_from_template_file`. That is not ideal, as users of the API may want to have access to the fields
defined in the template, before converting the template into builders.
Now the API user can load the template into memory via `CircuitCreateTemplate::from_yaml_file`
and have access to its fields as needed before converting it into builders via `CircuitCreateTemplate::into_builders`.
